AxiomLib: добавлены метрики Dice и Jaccard.
git-svn-id: https://svn.lvk.cs.msu.su/axiomlib/branches/armkor@158 08e97c73-d1ae-4023-87e5-8c5be0462271

+++ src/MultiMarking/MetricsEuclidean.cxx
@@ -23,7 +23,7 @@ namespace MultiMarking {
             }
         }
         s=sqrt(s);
-        s=s / sqrt(v1.size()) ;
+        s=s / sqrt((double)v1.size()) ;
         return s;
     }
